fix(workflows): register code workflows in CLI/worker bootstrap (#4257)
The in-memory code-workflow registry was populated only by the Next.js
app's bootstrap.ts, so the separate 'mercato workers' process (and any
CLI command) ran with an empty registry. Any instance started from an
unpersisted code-defined workflow (e.g. the checkout demo on a fresh
install) could not be resumed after an async activity: the activity
worker failed with 'Workflow definition not found' and the instance
stayed WAITING_FOR_ACTIVITIES forever. Reproduced on both develop and
current main.

Moves the registry store to @open-mercato/shared (same pattern as the
search module configs), keeps the validating wrapper and all existing
exports at the stable @open-mercato/core/modules/workflows/lib/code-registry
path, loads workflows.generated.ts in bootstrapFromAppRoot with a
fail-open fallback, and registers the definitions in the bootstrap
factory so every runtime shares one registry.
